Move common logic out of jest to be reused for jasmine
